logic corrected of passing userid with todo

+++ src/app/api/todos/route.ts
@@ -1,5 +1,6 @@
 import { connect } from '@/config/db';
 import Todo from '@/models/todo';
+import { getUserId } from '@/utils';
 import { NextRequest, NextResponse } from 'next/server';
 
 connect();
@@ -7,7 +8,7 @@ connect();
 
 export async function GET(req: NextRequest) {
     try {
-        const todos = await Todo.find({});
+        const todos = await Todo.find({createdBy: await getUserId(req)});
         return NextResponse.json({ success: true, data: todos }, { status: 200 });
       } catch (error) {
         return NextResponse.json({ success: false }, { status: 400 });
@@ -19,10 +20,13 @@ export async function GET(req: NextRequest) {
 export async function POST(req: NextRequest) {
     try {
         const { title, description = '' } = await req.json(); 
-        const todo = new Todo({title, description});
+        const userId = await getUserId(req);
+        console.log({userId});
+        const todo = new Todo({title, description, createdBy: userId});
         await todo.save();
         return NextResponse.json({ success: true, data: todo }, { status: 201 });
       } catch (error) {
+        console.error(error);
         return NextResponse.json({ success: false }, { status: 400 });
       }  
 }
@@ -30,6 +34,10 @@ export async function POST(req: NextRequest) {
 export async function PUT(req: NextRequest) {
     try {
         const reqBody = await req.json();
+        const { createdBy } =  await Todo.findById(reqBody._id);
+        if (createdBy.toString() !== await getUserId(req)) {
+          throw new Error('Unauthorized');
+        }
         const todo = await Todo.findByIdAndUpdate(reqBody._id, reqBody, {
           new: true,
           runValidators: true,
@@ -46,7 +54,8 @@ export async function PUT(req: NextRequest) {
 export async function DELETE(req: NextRequest) {
     try {
         const { _id} = await req.json();
-        const deletedTodo = await Todo.deleteOne({ _id });
+        const userId = await getUserId(req);
+        const deletedTodo = await Todo.deleteOne({ _id, createdBy: userId });
         if (!deletedTodo) {
           return NextResponse.json({ success: false }, {status: 404});
         }
@@ -56,4 +65,3 @@ export async function DELETE(req: NextRequest) {
         return NextResponse.json({ success: false }, { status: 400 });
       }
 }
-
